Collections working with user types

+++ src/response.hpp
@@ -83,7 +83,7 @@ private:
 private:
   uint8_t version_;
   int8_t flags_;
-  int8_t stream_;
+  int16_t stream_;
   uint8_t opcode_;
   int32_t length_;
   size_t received_;
